Implement __bool__ for diskcache.Cache to make it Truthy (#1)
Python first checks __bool__ and then __len__ to determine whether an object is Truthy. Given the fact that an empty cache returns 0 for the length, it breaks checks that rely on `if cache` instead of `if cache is not None` to detect whether it's initialized. This commit fixes the issue.